extracted ProfileButton from AppFrame
